Features/sort data impact usage field

+++ src/Databases/WorkflowDatabase.EF/Models/HpdUsage.cs
@@ -4,6 +4,6 @@
     {
         public int HpdUsageId { get; set; }
         public string Name { get; set; }
-        public int SortIndex { get; set; }
+        public byte SortIndex { get; set; }
     }
 }
